Refactor multi into separate files (#3209)
* splitting the MULTI into separate hpp/cpp files per class, to simplify further dev

* scrubbing redundant include files
